App crashes – Move `BloggingPrompt` to WordPressData
