+++ WordPress/Classes/Extensions/Post+BloggingPrompts.swift
@@ -1,4 +1,5 @@
 import Foundation
+import WordPressData
 
 extension Post {
