+++ WordPress/Classes/ViewRelated/Blog/Blog Dashboard/Cards/Prompts/DashboardPromptsCardCell.swift
@@ -1,5 +1,6 @@
 import UIKit
 import SafariServices
+import WordPressData
 import WordPressShared
 import WordPressFlux
