+++ WordPress/Classes/ViewRelated/Blog/Blogging Prompts/BloggingPromptCoordinator.swift
@@ -1,4 +1,5 @@
 import UIKit
+import WordPressData
 
 /// Helps manage the flow related to Blogging Prompts.
 ///
