+++ WordPress/Classes/ViewRelated/Blog/Blogging Prompts/BloggingPromptTableViewCell.swift
@@ -1,3 +1,5 @@
+import WordPressData
+
 class BloggingPromptTableViewCell: UITableViewCell, NibReusable {
 
     // MARK: - Properties
