+++ WordPress/Classes/ViewRelated/Blog/Blogging Prompts/BloggingPromptsViewController.swift
@@ -1,4 +1,5 @@
 import UIKit
+import WordPressData
 
 class BloggingPromptsViewController: UIViewController, NoResultsViewHost {
