+++ WordPress/Classes/ViewRelated/Blog/Blogging Prompts/RootViewCoordinator+BloggingPrompt.swift
@@ -1,3 +1,4 @@
+import WordPressData
 
 /// Encapsulates logic related to Blogging Prompts in RootViewCoordinator.
 ///
