+++ WordPress/Classes/ViewRelated/Feature Introduction/Blogging Prompts/BloggingPromptsIntroductionPresenter.swift
@@ -1,5 +1,6 @@
 import Foundation
 import UIKit
+import WordPressData
 
 /// Presents the BloggingPromptsFeatureIntroduction with actionable buttons
 /// and directs the flow according to which action button is tapped.
