+++ WordPress/Classes/ViewRelated/Post/EditPostViewController.swift
@@ -1,4 +1,5 @@
 import UIKit
+import WordPressData
 
 class EditPostViewController: UIViewController {
