+++ WordPress/Classes/ViewRelated/System/Floating Create Button/CreateButtonCoordinator.swift
@@ -1,4 +1,5 @@
 import Gridicons
+import WordPressData
 import WordPressFlux
 import WordPressUI
